% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_individual_path.R
\name{simulate_individual_path}
\alias{simulate_individual_path}
\title{function to get the path of 10000 individuals}
\usage{
simulate_individual_path(
  init_age,
  init_state,
  params,
  gender,
  i,
  cohort = 10000,
  model
)
}
\arguments{
\item{init_age}{the initial age of the path}

\item{init_state}{the initial state of all individuals}

\item{params}{the input parameters from the estimation results, read from an excel file}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index}

\item{cohort}{the number of simulations, set to be 10000}

\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}
}
\value{
10000 rows of individual paths in the states, 0 for H, 1 for M, 2 for D, 3 for MD, -1 for Dead
for each row it starts from the initial age as an input, and end at age 110
}
\description{
function to get the path of 10000 individuals
}
\examples{
simulated_individual_path=simulate_individual_path(init_age=65, init_state=0, params, gender, i, cohort = 10000, model=3)
}
