% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_life_table.R
\name{simulate_life_table}
\alias{simulate_life_table}
\title{the function to get n_sim number of lifetables}
\usage{
simulate_life_table(params, age, gender, i, latent, n_sim = 100, model)
}
\arguments{
\item{params}{the input parameters from the estimation results, read from an excel file}

\item{age}{initial age of the lifetable}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{n_sim}{the number of simulations}

\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}
}
\value{
a list of n_sum number of life table matrices
for each matrix, the row represents the age from the input initial age to 110, and the columns are states H M D MD Dead
for model 3 the frailty model, it simulates the latent factor to get n_sim number of lifetables, so we can get a distribution of the elements in the lifetable
for model 1 and 2, n_sim is suggest to set to be 1 to get one lifetable, otherwise it will produce the same lifetable n_sim times
}
\description{
the function to get n_sim number of lifetables
}
\examples{
simulated_lifetable=simulate_life_table(params,age,gender,i,latent,n_sim=100, model=3)
}
