% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_probability_5_frailty.R
\name{transition_probability_5_frailty}
\alias{transition_probability_5_frailty}
\title{function to get a matrix of transition probabilities at a certain age}
\usage{
transition_probability_5_frailty(params, age, gender, i, latent, model)
}
\arguments{
\item{params}{the input parameters from the estimation results, read from an excel file}

\item{age}{age of the individual}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}
}
\value{
5 times 5 matrix of transitions probabilities, the states are H M D MD Dead on the rows and columns
}
\description{
function to get a matrix of transition probabilities at a certain age
}
\examples{
transition_probabilities=transition_probability_5_frailty(params, init_age, gender, i, latent, model=3)
}
