% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_rate_5_frailty.R
\name{transition_rate_5_frailty}
\alias{transition_rate_5_frailty}
\title{function to calculate transition rates at a certain age}
\usage{
transition_rate_5_frailty(params, age, gender, i, latent, model)
}
\arguments{
\item{params}{the input parameters from the estimation results, read from an excel file}

\item{age}{age of the individual}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}
}
\value{
12 times 1 vector of transition rates for the 12 types of transitions
}
\description{
function to calculate transition rates at a certain age
}
\examples{
transition_rates=transition_rate_5_frailty(params, init_age, gender, i, latent, model=3)
}
